requirements.txt
fastmcp
langchain-mcp-adapters
langchain
langchain-openai
psycopg2-binary
pandas
python-dotenv
sqlalchemy
langchain-chroma
chromadb